fixed ApplicationException.

+++ application/views.py
@@ -1,8 +1,6 @@
-from django import http
-from django.template import loader, RequestContext, Context
 from django.template.response import TemplateResponse
 from application.models import *
-from application.exceptions import Http400
+from application.exceptions import Http404
 
 
 # -----------------------------------------------
@@ -63,50 +61,31 @@ def delete_post(request, post_id):
         key=post_id
     )
     post = view.first()
+    if post is None:
+        raise Http404
     post.delete()
 
     return HttpResponse(status=200)
 
-
-# -----------------------------------------------
-# error views
-# -----------------------------------------------
-def bad_request(request):
-    template = loader.get_template('error/default.html')
-    model = ErrorViewModel(
-        status=400,
-        exception='Bad Request'
+def update_post(request, post_id):
+    """
+    Update the post.
+    :param post_id: The post id.
+    :param request.body:
+        It should be json.
+        title: The post title.
+        content The post content.
+    :return: JsonResponse(PostModel)
+    """
+    model = json.loads(request.body)
+    view = PostModel.view(
+        'posts/all',
+        key=post_id
     )
-    return http.HttpResponseBadRequest(template.render(RequestContext(request, model)))
-
-def permission_denied(request):
-    template = loader.get_template('error/default.html')
-    model = ErrorViewModel(
-        status=403,
-        exception='Permission Denied'
-    )
-    return http.HttpResponseForbidden(template.render(RequestContext(request, model)))
-
-def page_not_found(request):
-    template = loader.get_template('error/default.html')
-    model = ErrorViewModel(
-        status=404,
-        exception='%s Not Found' % request.path
-    )
-    return http.HttpResponseNotFound(template.render(RequestContext(request, model)))
-
-def method_not_allowed(request):
-    template = loader.get_template('error/default.html')
-    model = ErrorViewModel(
-        status=405,
-        exception='%s Not Allowed' % request.method
-    )
-    return http.HttpResponse(status=405, content=template.render(RequestContext(request, model)))
-
-def server_error(request):
-    template = loader.get_template('error/default.html')
-    model = ErrorViewModel(
-        status=500,
-        exception='Server Error'
-    )
-    return http.HttpResponseServerError(template.render(Context(model)))
+    post = view.first()
+    if post is None:
+        raise Http404
+    post.title = model.get('title')
+    post.content = model.get('content')
+    post.save()
+    return JsonResponse(post.dict())
